Change helper layout
* Only use platform specific code for drive operations.
* Break windows and mac builds.
* Make linux code obey the new layout.
* Add mount option.
