+++ helper/linux/linux.pri
@@ -1,0 +1,10 @@
+INCLUDEPATH += $$top_srcdir/helper/linux/
+DESTDIR = $$top_srcdir/helper/linux/
+
+QT += dbus
+
+CONFIG += link_pkgconfig
+PKGCONFIG += liblzma
+
+SOURCES += linux/drive.cpp
+HEADERS += linux/drive.h
